Fix issues with valid cards
Heartbreak no longer forgets its previous true status until the new
hand and follow_suit was rewritten to make the logic more sound.

Additionally, a stub for a TestGame class was created.
This class is used so that you can give a predefined deal and play one
hand in a deterministic manner.